python_requirements.txt
numpy==1.23.5
pandas==1.5.3
scipy==1.11.1
scikit-learn==1.4.2
matplotlib==3.6.3
click==8.1.3
numba==0.57.0
optuna==4.0.0
lightgbm==4.5.0
pyod==2.0.2
nemea-fet==0.1.3
torch==2.2.1